refactor(lib): migrate components to hooks
