+++ lib/widgets.dart
@@ -1,5 +1,6 @@
 library widgets;
 
 export 'src/widgets/query_wizard.dart';
+export 'src/widgets/query_wizard_view.dart';
 export 'src/widgets/tabs/tables_and_fields.dart';
 export 'src/widgets/tabs/parts/sources_tree.dart';
